requirements.txt
websocket-client>=0.35.0